Consolidate turn analytic registrations into registry.py.
Aggregate tuple, derived catalog, catalog_entry lookup, and dispatch now
live in one module; registrations.py is removed.
