+++ packages/bff/bff/analytics/registry.py
@@ -1,7 +1,7 @@
 """BFF analytics catalog and dispatch."""
 
 from api.analytics.catalog import tuple_aligned_with_turn_analytic_catalog
-from api.analytics.registrations import TURN_ANALYTIC_CATALOG
+from api.analytics.registry import TURN_ANALYTIC_CATALOG
 from api.diagnostics import Diagnostics
 
 from bff.analytics.descriptor import AnalyticDescriptor
